separating routing logic from display/fetch
